backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
google-generativeai>=0.8.0
fastembed>=0.7.0
qdrant-client==1.12.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18
markdown==3.7
beautifulsoup4==4.12.3
httpx==0.28.1
